More cool stuff for my personal website
Adding in some Reddit, some games, and of course some good old common CS induction problems. 

This website is starting to be a powerhouse and is really telling a life story about my time at BU. 

If you want to be challenged / get ready for interview questions, attempt the proofs of induction on the PDF file.